refactor: follow naming convention in RTK store
