+++ src/store/store.tsx
@@ -1,19 +1,19 @@
 import { configureStore } from '@reduxjs/toolkit';
-import authSlice from './Slices/users/slice';
-import attractionsSlice from './Slices/attractions/slice';
-import privacySlice from './Slices/users/privacySlice';
-import profileSlice from './Slices/users/profileSlice';
-import activitiesReducer from './Slices/activitiesSlice';
+import authReducer from './Slices/users/slice';
+import attractionsReducer from './Slices/attractions/slice';
+import privacyReducer from './Slices/users/privacySlice';
+import profileReducer from './Slices/users/profileSlice';
+import activitiesReducer from './Slices/activities/slice';
 import plansReducer from './Slices/plans/slice';
 import reviewsReducer from './Slices/reviews/slice';
 import logsReducer from './Slices/paymentLogs/slice';
 
 const store = configureStore({
   reducer: {
-    auth: authSlice,
-    attractionsThunk: attractionsSlice,
-    privacy: privacySlice,
-    profile: profileSlice,
+    auth: authReducer,
+    attractionsThunk: attractionsReducer,
+    privacy: privacyReducer,
+    profile: profileReducer,
     activities: activitiesReducer,
     plans: plansReducer,
     reviews: reviewsReducer,
